Added option to use a roaming config file, so it stays for each version if people don't want to copy their settings from a previous version.
This, however, has quite a lot of bad code and probably has issues yet to be found.

First: The code is dirty, it's not finished, and there should definitely NOT be a release with this code, it should be cleaned up, tested more, and fixed more before being released.

Secondly: Because of laziness, it's using C:\Users\Public, instead of appdata, because of issues I ran into that I wont explain right now because I'm just glad this is mostly working, because I've been working on it for 1.5-2.5 hours.

Read some of the comments in the code if you want more information.