My untested workaround for the FONT (icons.ttf) issue.
Remove the registry reference to the FONT.
Move the file inside the DDU directory and rename it, and try to remove it(it will most of the time fail, because it is in use, that's the main issue)
Continue removal of the AMD folders (will work now as the problematic file has been moved.)

Idea has been tested but the code here is untested. Will do when I have more time.

Todo: find the real way to remove the font/file.  (the AMD installer is able to do so)