Removed paexec.exe
DDU is now impersonate the Windows system account by "stealing" with OpenProcessToken and using ImpersonateLoggedOnUser instead of starting directly as a SYSTEM process.

This will help with Antivirus with false positive because of PAEXEC.

Note: this doesn't work on all SYSTEM process. An always working one seem to be LSASS.